requirements.txt
msgpack==1.0.8
PyNaCl==1.5.0
click==8.1.7
uvloop==0.20.0; platform_system != "Windows"